CHECK: remove TODO
Not sure how to implement argument unpacking for the $this->set_cache() call as there is a fourth parameter that is not part of `$cache_keys` named `$in_nonce_check`. I need to check if I'm not missing something, and it is indeed ok to simply remove the TODO comment.