PR 2619 DB/DirectDatabaseQuery: fix handling of namespaced calls to cache functions
Before this sniff was not properly differentiating between fully qualified calls to global WP cache functions, and namespaced calls to non-WP functions with the same name as the WP functions.

In other words, the sniff was correctly identifying `\wp_cache_get()` as a WP cache function, but it was incorrectly identifying `MyNamespace\wp_cache_get()` as a WP cache function. This commit fixes this issue.