WP/GlobalVariablesOverride: update for PHPCS 4.0
The tokenization of (namespaced) "names" has changed in PHP 8.0, and this new tokenization will come into effect for PHP_CodeSniffer as of version 4.0.0.

This commit adds handling for the new tokenization of fully qualified function/method calls when checking if the sniff can determine the key used to retrieve a value from the `$GLOBALS` array. The sniff was already handling this correct by chance due to a check that happens later in the code execution to bow out if `$var_name` is empty.