WP/CronInterval: update for PHPCS 4.0
The tokenization of (namespaced) "names" has changed in PHP 8.0, and this new tokenization will come into effect for PHP_CodeSniffer as of version 4.0.0.

This commit adds handling for the new tokenization when checking if a WP time constant was used and when checking the name of a first-class callable. It also updates a few test expectation to account for the difference in the sniff behavior when running with PHPCS 3.x and 4.x.