WIP ContextHelper::is_in_function_call(): add basic tests
Check the way the tests are structured. There is for sure room to improve their organization.

Also check the way that I'm running the tests in the WordPress/Util/Tests in composer.json and the need to add a custom autoloader to bootstrap.php. There is probably a better way to achieve both.