CHECK Security/EscapeOutput: update for PHPCS 4.0
The tokenization of (namespaced) "names" has changed in PHP 8.0, and this new tokenization will come into effect for PHP_CodeSniffer as of version 4.0.0.

This commit adds handling for the new tokenization to this sniff.

Discuss if it is worth supporting `\true` and `\false` (and maybe even `\null`) in this sniff.