WP/EnqueuedResourceParameters: remove PHPCS 3.x code
Now that we are dropping support for PHPCS 3.x, this sniff can be updated to remove code that was specific to this version. Starting with PHPCS 4.0, `\false` is tokenized as `T_FALSE` (before it was tokenized as `T_NS_SEPARATOR` + `T_FALSE`), so it is not necessary anymore to have `T_NS_SEPARATOR` in `EnqueuedResourceParametersSniff::$false_tokens`.